requirements.txt
langchain
langchain_community
bs4
sentence-transformers
chromadb
flask